refactor: rename package paths for FilterContent and SnackBarState
